+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.3.0 (in development)
+-----------------------
+- A session-wide `Accept` header can now be specified when constructing a
+  `Client` without having to use a custom `Session`
+
 v0.2.0 (2023-11-03)
 -------------------
 - The `version` argument to `make_user_agent()` is now optional